Move loadAppFonts into own file
