+++ lib/spot.dart
@@ -1,12 +1,7 @@
 /// Spot is a library for testing the Widget tree of Flutter apps.
 library spot;
 
-import 'dart:convert';
-import 'dart:io';
-
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
-import 'package:flutter_test/flutter_test.dart';
 import 'package:spot/spot.dart';
 import 'package:spot/src/spot/selectors.dart' show Spot;
 import 'package:spot/src/spot/widget_selector.dart';
@@ -18,9 +13,11 @@ export 'package:checks/checks.dart'
         // async checks are not useful for spot
         StreamChecks,
         WithQueueExtension;
-export 'package:checks/context.dart' show Condition, Context, ContextExtension, Extracted, Rejection, Subject;
+export 'package:checks/context.dart'
+    show Condition, Context, ContextExtension, Extracted, Rejection, Subject;
 export 'package:meta/meta.dart' show useResult;
 export 'package:spot/src/act/act.dart' show Act, act;
+export 'package:spot/src/screenshot/load_app_fonts.dart' show loadAppFonts;
 export 'package:spot/src/screenshot/screenshot.dart'
     show
         ElementScreenshotExtension,
@@ -28,15 +25,20 @@ export 'package:spot/src/screenshot/screenshot.dart'
         SelectorScreenshotExtension,
         SnapshotScreenshotExtension,
         takeScreenshot;
-export 'package:spot/src/spot/default_selectors.dart' show DefaultWidgetMatchers, DefaultWidgetSelectors;
-export 'package:spot/src/spot/diagnostic_props.dart' show DiagnosticPropWidgetMatcher, DiagnosticPropWidgetSelector;
+export 'package:spot/src/spot/default_selectors.dart'
+    show DefaultWidgetMatchers, DefaultWidgetSelectors;
+export 'package:spot/src/spot/diagnostic_props.dart'
+    show DiagnosticPropWidgetMatcher, DiagnosticPropWidgetSelector;
 export 'package:spot/src/spot/effective/effective_text.dart'
     show EffectiveTextMatcher, EffectiveTextSelector, TextStyleSubject;
 export 'package:spot/src/spot/filters/child_filter.dart' show ChildFilter;
 export 'package:spot/src/spot/filters/parent_filter.dart' show ParentFilter;
-export 'package:spot/src/spot/filters/predicate_filter.dart' show PredicateFilter;
-export 'package:spot/src/spot/filters/widget_type_filter.dart' show WidgetTypeFilter;
-export 'package:spot/src/spot/finder_interop.dart' show FinderFilter, FinderToSpot, SpotToFinder;
+export 'package:spot/src/spot/filters/predicate_filter.dart'
+    show PredicateFilter;
+export 'package:spot/src/spot/filters/widget_type_filter.dart'
+    show WidgetTypeFilter;
+export 'package:spot/src/spot/finder_interop.dart'
+    show FinderFilter, FinderToSpot, SpotToFinder;
 export 'package:spot/src/spot/matcher_generator.dart' show CreateMatchers;
 export 'package:spot/src/spot/props.dart'
     show
@@ -70,9 +72,15 @@ export 'package:spot/src/spot/snapshot.dart'
         WidgetSnapshot,
         WidgetSnapshotShorthands;
 export 'package:spot/src/spot/text/any_text.dart' show AnyText;
-export 'package:spot/src/spot/tree_snapshot.dart' show ScopedWidgetTreeSnapshot, WidgetTreeNode;
+export 'package:spot/src/spot/tree_snapshot.dart'
+    show ScopedWidgetTreeSnapshot, WidgetTreeNode;
 export 'package:spot/src/spot/widget_matcher.dart'
-    show MultiWidgetMatcher, MultiWidgetMatcherExtensions, PropertyCheckFailure, WidgetMatcher, WidgetMatcherExtensions;
+    show
+        MultiWidgetMatcher,
+        MultiWidgetMatcherExtensions,
+        PropertyCheckFailure,
+        WidgetMatcher,
+        WidgetMatcherExtensions;
 export 'package:spot/src/spot/widget_selector.dart'
     show
         ElementFilter,
@@ -483,84 +491,3 @@ WidgetSelector<W> spotKeys<W extends Widget>(
     children: children,
   );
 }
-
-/// Loads the app fonts so they are readable in tests.
-/// This is especially helpful when using takeScreenshot();
-Future<void> loadAppFonts() async {
-  TestWidgetsFlutterBinding.ensureInitialized();
-  // Load the font manifest
-  final fontManifest = await rootBundle.loadStructuredData<List<Map<String, dynamic>>>(
-    'FontManifest.json',
-    (string) async {
-      // Parse and cast the JSON to List<Map<String, dynamic>>
-      return (json.decode(string) as List).map((e) => e as Map<String, dynamic>).toList();
-    },
-  );
-  await loadFontsFromFontsManifest(fontManifest);
-}
-
-@visibleForTesting
-Future<void> loadFontsFromFontsManifest(List<Map<String, dynamic>> fontManifest) async {
-  // When no special fonts are defined or there is no roboto in the manifest we load it from the manifest to render standard Text widgets
-  if (fontManifest.isEmpty || fontManifest.every((element) => element['family'] != 'Roboto')) {
-    final fontLoader = FontLoader('Roboto');
-    final fontPath = Platform.executable;
-    final desiredPath = fontPath.split('/engine/')[0];
-    fontLoader.addFont(rootBundle.load('$desiredPath/material_fonts/Roboto-Regular.ttf'));
-    await fontLoader.load();
-  }
-
-  // Load the fonts from the manifest. These can be the ones defined in the app or in dependencies.
-  for (final Map<String, dynamic> font in fontManifest) {
-    final fontLoader = FontLoader(derivedFontFamily(font));
-    // Cast 'fonts' field to List<Map<String, dynamic>> to use in the for loop
-    final List<Map<String, dynamic>> fontsList = (font['fonts'] as List).map((e) => e as Map<String, dynamic>).toList();
-
-    for (final Map<String, dynamic> fontType in fontsList) {
-      final String assetPath = fontType['asset'] as String;
-      print('assetPath: $assetPath');
-      fontLoader.addFont(rootBundle.load(assetPath));
-    }
-
-    await fontLoader.load();
-  }
-}
-
-@visibleForTesting
-String derivedFontFamily(Map<String, dynamic> fontDefinition) {
-  if (!fontDefinition.containsKey('family')) {
-    return '';
-  }
-  print('fontDefinition: $fontDefinition');
-
-  final String fontFamily = fontDefinition['family'] as String;
-
-  if (_overridableFonts.contains(fontFamily)) {
-    return fontFamily;
-  }
-
-  if (fontFamily.startsWith('packages/')) {
-    final fontFamilyName = fontFamily.split('/').last;
-    if (_overridableFonts.any((font) => font == fontFamilyName)) {
-      return fontFamilyName;
-    }
-  } else {
-    for (final fontType in fontDefinition['fonts'] as List<dynamic>) {
-      print('fontType: $fontType');
-      final String? asset = fontType['asset'] as String?;
-      if (asset != null && asset.startsWith('packages')) {
-        final packageName = asset.split('/')[1];
-        return 'packages/$packageName/$fontFamily';
-      }
-    }
-  }
-  return fontFamily;
-}
-
-const List<String> _overridableFonts = [
-  'Roboto',
-  '.SF UI Display',
-  '.SF UI Text',
-  '.SF Pro Text',
-  '.SF Pro Display',
-];
